+++ packages/cli/src/feManifest.ts
@@ -1,0 +1,5 @@
+const config = {};
+
+export { config };
+const manifest = { config };
+export default manifest;