+++ packages/cli/src/generator_templates/js/emails/components/Header.jsx
@@ -1,15 +1,15 @@
 import { MjmlSection, MjmlColumn, MjmlImage } from "mjml-react";
 
-const Header = ({ big }) => {
+const Header = ({ loose }) => {
   return (
-    <MjmlSection padding={big ? "48px 0 40px" : "48px 0 24px"}>
+    <MjmlSection padding={loose ? "48px 0 40px" : "48px 0 24px"}>
       <MjmlColumn>
         <MjmlImage
           padding="0 24px 0"
-          width={big ? "146px" : "91px"}
-          height={big ? "32px" : "20px"}
-          align="left"
-          src="https://s3.amazonaws.com/lab.campsh.com/mailing-logo%402x.png"
+          width="49px"
+          height="54px"
+          align="center"
+          src="https://s3.amazonaws.com/lab.campsh.com/bb-transparent%402x.png"
           cssClass="logo"
         />
       </MjmlColumn>