+++ packages/cli/src/generator_templates/ts/emails/components/ButtonPrimary.tsx
@@ -1,5 +1,5 @@
 import { MjmlButton } from "mjml-react";
-import { black, grayLight } from "./theme";
+import { black, gold, grayLight } from "./theme";
 import { leadingTight, textBase, borderBase } from "./theme";
 
 type ButtonPrimaryProps = {
@@ -13,11 +13,12 @@ const ButtonPrimary: React.FC<ButtonPrimaryProps> = ({ link, uiText }) => {
       <MjmlButton
         lineHeight={leadingTight}
         fontSize={textBase}
-        height={52}
+        height={32}
         padding="0"
         align="left"
         href={link}
         backgroundColor={black}
+        color={grayLight}
         borderRadius={borderBase}
         cssClass="light-mode"
       >
@@ -26,11 +27,11 @@ const ButtonPrimary: React.FC<ButtonPrimaryProps> = ({ link, uiText }) => {
       <MjmlButton
         lineHeight={leadingTight}
         fontSize={textBase}
-        height={52}
+        height={32}
         padding="0"
         align="left"
         href={link}
-        backgroundColor={grayLight}
+        backgroundColor={gold}
         color={black}
         borderRadius={borderBase}
         cssClass="dark-mode"