+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # mailing
 
+## 0.7.3
+
+### Patch Changes
+
+- 2b666f0: switch fs.watch to chokidar.watch
+- Updated dependencies [2b666f0]
+  - mailing-core@0.7.3
+
+## 0.7.2
+
+### Patch Changes
+
+- c87cc5b: remove an additional / from api route (fix #158)
+- 3e324e8: - Introduce esbuild for bundling templates to put in .mailing. Previously, we copied the templates, but bundling greatly improves the ability to include modules and resolve paths.
+  - Build a feManifest.js that lets the frontend app get the user’s config settings
+  - Update tsconfigs to support above changes
+  - Fix a bad bug where moduleManifest.js always set templates to `{}` because the function invoked by `filter` had no return value.
+- Updated dependencies [c87cc5b]
+- Updated dependencies [3e324e8]
+  - mailing-core@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes